requirements.txt
ccxt==4.5.44
python-dotenv==1.1.0
structlog==25.1.0
tenacity==9.1.4
pyyaml==6.0.2
pytest==8.3.5
pytest-mock==3.14.0
pytest-cov==6.1.0
